Finance Management Tool/ml-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0
joblib==1.3.2
python-dateutil==2.8.2
dateparser==1.2.0
